html/browsers/origin/origin-keyed-agent-clusters/regression-1399759.https.sub.html

<!DOCTYPE html>
<html>
<head>
<title>Agent Equality After Navigating about:blank</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
</head>
<body>
<script>
// Regression test for crbug.com/1399759. This is mainly based on
// html/infrastructure/urls/terminology-0/document-base-url-initiated-grand-parent.https.window.html,
// but restricts itself to the exact error condition.
onload = () => {
  promise_test(async test => {
    // Create a cross-origin iframe. Use the executor.html, so we can ask it
    // to execute scripts for us.
    const child_token = token();
    const iframe = document.createElement("iframe");
    iframe.src = get_host_info().HTTPS_REMOTE_ORIGIN +
      `/common/dispatcher/executor.html?uuid=${child_token}`;
    document.body.appendChild(iframe);

    // The child creates a grand child in an iframe.
    const reply_token = token();
    send(child_token, `
      const iframe = document.createElement("iframe");
      iframe.src = "/common/blank.html";
      iframe.onload = () => {
        send("${reply_token}", "grand child loaded");
      };
      document.body.appendChild(iframe);
    `);
    assert_equals(await receive(reply_token), "grand child loaded");

    // Navigate the grand-child toward about:blank.
    // Navigation are always asynchronous. There is no way to to know when
    // about:blank document committed. A timer is used instead:
    const grandchild = iframe.contentWindow[0];
    grandchild.location = "about:blank";
    await new Promise(r => test.step_timeout(r, /*ms=*/500));

    // This document and grandchild are same-origin, because about:blank
    // inherits its origin from the initiator of the navigation, which is us.
    // This access should not throw.
    grandchild.document;
  }, "Check the baseURL of an about:blank document cross-origin with its parent");
};

</script>
</body>
</html>
